third_party/WebKit/LayoutTests/geolocation-api/enabled.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests that the navigator.geolocation object is present.");

function hasGeolocationProperty()
{
    for (var property in navigator) {
        if (property == "geolocation")
            return true;
    }
    return false;
}

shouldBeTrue("typeof navigator.geolocation == 'object'");
shouldBeTrue("hasGeolocationProperty()");
shouldBeTrue("'geolocation' in navigator");

window.jsTestIsAsync = false;
</script>
</body>
</html>
